dev_requirements.txt
asn1crypto==0.24.0
boto3==1.9.17
botocore==1.12.17
cffi==1.11.5
cryptography==2.3.1
docutils==0.14
idna==2.7
jmespath==0.9.3
numpy==1.15.2
pandas==0.23.4
psycopg2-binary==2.7.3.2
pycparser==2.19
PyMySQL==0.9.2
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.12
s3transfer==0.1.13
six==1.11.0
urllib3==1.23
